Hospital_Management_System/requirements.txt
Django>=4.2,<5.0
django-jazzmin>=3.0,<4.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
reportlab>=4.2,<5.0
